services/main_service/pyproject.toml
[project]
name = "main_service"
version = "0.1.0"
description = "Public BFF: proxies auth, handles uploads, lists videos & notifications"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1",
    "db_core",
    "fastapi[standard]>=0.115",
    "httpx>=0.27",
    "pydantic-settings>=2.5",
    "pyjwt>=2.9",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.9",
    "uvicorn>=0.30",
]

[tool.uv.sources]
db_core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "httpx>=0.27",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
